last digit

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -9,18 +9,18 @@
 int main(void)
 {
 	int n;
-	int last;
+	long int last;
 
 	srand(time(0));
 	n = rand() - RAND_MAX / 2;
 	/* your code goes there */
 	last = n % 10;
 	if (last > 5)
-		printf("Last digit of %d is %d and is greater than 5\n", &n, &last);
+		printf("Last digit of %d is %ld and is greater than 5\n", &n, &last);
 	if (last == 0)
-		printf("Last digit of %d is %d and is 0\n", &n, &last);
+		printf("Last digit of %d is %ld and is 0\n", &n, &last);
 	if (6 < last != 0)
-		print("Last digit of %d is %d and is less than 6 but not 0\n", &n, &last);
+		print("Last digit of %d is %ld and is less than 6 but not 0\n", &n, &last);
 
 	return (0);
 }
